requirements.txt
matrix-bot-api == 0.1
PyYAML ~= 3.0